refactor datamodel's to own subpackage
